+++ src/shared/line.h
@@ -10,7 +10,7 @@ typedef struct line {
 /**
  * Creates a new line with the given coordinates.
  */
-line* new_line(double x0, double y0, double x1, double y1);
+line* new_line(float x0, float y0, float x1, float y1);
 
 /**
  * Frees the line 'ln.
